Move ServiceFactory to remote package
